% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccspectral.df.R
\name{ccspectral.df}
\alias{ccspectral.df}
\title{function to segmentate an image between plant and soil using autothreshold,
supervised autothreshold or manual criteria. Measurement of plant
physiological activity.}
\usage{
ccspectral.df(
  tif.path,
  chart,
  obs.areas,
  pdf = F,
  calculate.thresh = F,
  descrip = F,
  manual.mask.test = F,
  index. = c("NDVI", "SR", "MSAVI", "EVI", "CI", "BSCI", "BI", "NORR", "NORG", "NORB",
    "EXR", "EXG", "EXB", "EXGR", "CIVE", "VEG", "HUE", "SAT", "VAL"),
  threshold.method = NULL,
  threshold.vector = NULL,
  descriptors. = c("median", "mean", "sd", "min", "max", "diff.range")
)
}
\arguments{
\item{tif.path}{string. File path where you can find the image files.}

\item{chart}{SpatialPolygons. Color chart tile position defined by 24
features one by each color tile.}

\item{obs.areas}{list. List of binary rasters with the observation area
defined by 1 and 0 to non interest area.}

\item{pdf}{logical.If you want to obtain a pdf with index and segmented
result. Default = F}

\item{calculate.thresh}{logical.If you want to calculate threshold
atomaticaly in the observed area. Default = F}

\item{descrip}{logical. If you want to calculate statistical descriptors
over the segmented surface values. Default = F}

\item{manual.mask.test}{logical. If you want to conduct a quality test
confronting the segmentation results with ground true areas, if true  you
need to provide this segmentation masks. Default = F}

\item{index.}{character. character vector with the names of spectral indices
that you want to calculate. Possible values:
"NDVI" (Normalized Differential Vegetation Index), "SR" (Simple Ratio),
"MSAVI"(Second Modified Soil Adjusted Vegetation Index),
"EVI"(Enhanced Vegetation Index), "CI"(Crust Index),
"BSCI"(Biological Soil Crust Index), "BI"(Brightness Index),
"NORR"(Normalized Red), "NORG"(Normalized Green), "NORB"(Normalized Blue),
"EXR"(Excess Red), "EXG"(Excess Green),
"EXB"(Excess Blue), "EXGR"(Excess Green minus Excess Red), "CIVE"(Color Index
of Vegetation Extraction), "VEG"(Vegetative),
"HUE"(Dominant colour wavelength level), "SAT"(Saturation), "VAL"(Value).
Default = c("NDVI", "SR", "MSAVI", "EVI", "CI", "BSCI", "BI", "NORR", "NORG",
"NORB", "EXR", "EXG", "EXB", "EXGR", "CIVE", "VEG", "HUE", "SAT", "VAL")}

\item{threshold.method}{character. if \emph{calculate.thresh} = T define the
autothershold method to apply over the calculated spectral index raster.
It coulb be one of this values: "Huang", "IsoData", "IJDefault", "Li",
"Mean", "MinErrorI", "Moments", "Otsu", "Percentile",
"RenyiEntropy", "Shanbhag", "Triangle". This function is based on
\emph{autothresholdr} R package (Landini et al., 2017).
Otherwise, an adaptation to R of Auto Threshold plugin from ImageJ.
(https://imagej.net/Auto_Threshold).
Default = NULL}

\item{threshold.vector}{numeric. if \emph{calculate.thresh} = F define the
thershold values for defined \emph{index.} values.
Values must be provided in the same order as "index." values.
Default = NULL}

\item{descriptors.}{character. If descrip = T you provide a vector with the
descriptors to be calculated in over the segmented areas. Values could be.
"median","mean","sd","min","max" and "diff.range"(range amplitude).}
}
\value{
A Data frame with the calculated values for the each image in the
tif.path
}
\description{
function to segmentate an image between plant and soil using autothreshold,
supervised autothreshold or manual criteria. Measurement of plant
physiological activity.
}
\author{
Manuel Molina-Bustamante
}
