% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tif2raster.R
\name{tif2raster}
\alias{tif2raster}
\title{import tif file and set as raster}
\usage{
tif2raster(msk_name)
}
\arguments{
\item{x}{string. Path where you can find the tif file.}
}
\value{
A raster file with three layers.
}
\description{
import tif file and set as raster
}
\author{
Manuel Molina-Bustamante.
}
